chore: apply suggestions from code review
